requirements.txt
aiogram==3.7.0
aiosqlite==0.20.0
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.30.1
